Stopped health from decreasing while game is paused (player is immune), and refactored lab, pause and extractor minigame windows into windows folder
