+++ source/core/src/main/com/csse3200/game/windows/LabWindow.java
@@ -1,4 +1,4 @@
-package com.csse3200.game;
+package com.csse3200.game.windows;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
@@ -13,7 +13,6 @@ import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 
 import com.csse3200.game.components.PowerupType;
-import com.csse3200.game.entities.Entity;
 import com.csse3200.game.input.InputOverrideComponent;
 import com.csse3200.game.services.ServiceLocator;
