+++ source/core/src/main/com/csse3200/game/windows/PauseWindow.java
@@ -1,4 +1,4 @@
-package com.csse3200.game.components.pause;
+package com.csse3200.game.windows;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
